refactor: split describe_parameters.py into separate files for each tool
